+++ CHANGELOG.rst
@@ -8,6 +8,7 @@ CHANGELOG
 * bug-fix: Remove unnecessary dependency tensorflow
 * doc-fix: Change ``distribution`` to ``distributions``
 * bug-fix: Increase docker-compose http timeout and health check timeout to 120.
+* feature: Local Mode: Add support for intermediate output to a local directory.
 
 1.16.1.post1
 ============